Proof of concept: OpenRadioss + preCICE
use the -preCICE option of the build_script, and a dedicated engine binary will be built.

A file *1.cpl should be present, containing the differents options:
Here is an example of a *1.cpl for a solver that write the positions and get the forces at the interface.

/PRECICE/PARTICIPANT_NAME/SolverOne
/PRECICE/CONFIG_FILE/precice-config.xml
/PRECICE/MESH_NAME/MeshOne
/PRECICE/WRITE/Position
/PRECICE/READ/Force
/PRECICE/INTERFACE/22

The id of the coupling interface corresponds to a GRNOD id.
Note that only explicit coupling scheme are available, a lightweight checkpoint/restart strategy is needed for implicit coupling scheme.